refactor: collect types for databases
