+++ models/sqlite/address.go
@@ -10,6 +10,7 @@ import (
 
 	"gorm.io/gorm"
 
+	"github.com/filecoin-project/venus-messager/models/mtypes"
 	"github.com/filecoin-project/venus-messager/models/repo"
 	"github.com/filecoin-project/venus-messager/types"
 )
@@ -22,8 +23,8 @@ type sqliteAddress struct {
 	SelMsgNum         uint64      `gorm:"column:sel_msg_num;type:unsigned bigint;NOT NULL"`
 	State             types.State `gorm:"column:state;type:int;index;default:1"`
 	GasOverEstimation float64     `gorm:"column:gas_over_estimation;type:decimal(10,2);"`
-	MaxFee            types.Int   `gorm:"column:max_fee;type:varchar(256);"`
-	MaxFeeCap         types.Int   `gorm:"column:max_fee_cap;type:varchar(256);"`
+	MaxFee            mtypes.Int  `gorm:"column:max_fee;type:varchar(256);"`
+	MaxFeeCap         mtypes.Int  `gorm:"column:max_fee_cap;type:varchar(256);"`
 
 	IsDeleted int       `gorm:"column:is_deleted;index;default:-1;NOT NULL"` // 是否删除 1:是  -1:否
 	CreatedAt time.Time `gorm:"column:created_at;index;NOT NULL"`            // 创建时间
@@ -49,10 +50,10 @@ func FromAddress(addr *types.Address) *sqliteAddress {
 	}
 
 	if !addr.MaxFee.Nil() {
-		sqliteAddr.MaxFee = types.NewFromGo(addr.MaxFee.Int)
+		sqliteAddr.MaxFee = mtypes.NewFromGo(addr.MaxFee.Int)
 	}
 	if !addr.MaxFeeCap.Nil() {
-		sqliteAddr.MaxFeeCap = types.NewFromGo(addr.MaxFeeCap.Int)
+		sqliteAddr.MaxFeeCap = mtypes.NewFromGo(addr.MaxFeeCap.Int)
 	}
 
 	return sqliteAddr
@@ -185,10 +186,10 @@ func (s sqliteAddressRepo) UpdateFeeParams(ctx context.Context, addr address.Add
 		updateColumns["gas_over_estimation"] = gasOverEstimation
 	}
 	if !maxFee.Nil() {
-		updateColumns["max_fee"] = types.NewFromGo(maxFee.Int)
+		updateColumns["max_fee"] = mtypes.NewFromGo(maxFee.Int)
 	}
 	if !maxFeeCap.Nil() {
-		updateColumns["max_fee_cap"] = types.NewFromGo(maxFeeCap.Int)
+		updateColumns["max_fee_cap"] = mtypes.NewFromGo(maxFeeCap.Int)
 	}
 	updateColumns["updated_at"] = time.Now()
